+++ src/model/ula-message.ts
@@ -1,6 +1,5 @@
-"use strict";
 /*
- * Copyright 2019 Coöperatieve Rabobank U.A.
+ * Copyright 2020 Coöperatieve Rabobank U.A.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,36 +13,48 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-Object.defineProperty(exports, "__esModule", { value: true });
+
 /**
- * The ULA Message type is used to
+ * The UlaMessage type is used to
  * send messages (events) to ULA plugins.
  * The object is entirely dynamic, but it
  * does require a 'type' field, so the
  * plugins can recognize (or ignore) the
  * event.
  */
-class Message {
-    constructor(obj) {
-        if (!obj.type) {
-            throw new ReferenceError('Type field is missing');
-        }
-        this._obj = obj;
+export class UlaMessage {
+  private readonly _obj: any
+
+  constructor (obj: any) {
+    if (!obj.type) {
+      throw new ReferenceError('Type field is missing')
     }
-    /**
-     * The dynamic properties of the message
-     * @return any
-     */
-    get properties() {
-        return this._obj;
-    }
-    /**
-     * Converts a this object to a json string
-     * @return object
-     */
-    toJSON() {
-        return this._obj;
-    }
+
+    this._obj = obj
+  }
+
+  /**
+   * The dynamic properties of the message
+   * @return any
+   */
+  public get properties (): any {
+    return this._obj
+  }
+
+  /**
+   * Converts a this object to a json object
+   * @return object
+   */
+  public toJSON (): object {
+    return this._obj
+  }
+
 }
-exports.Message = Message;
-//# sourceMappingURL=message.js.map
+
+/**
+ * @deprecated Please use UlaMessage
+ * @see UlaMessage
+ */
+export class Message extends UlaMessage {
+
+}