+++ test/unit/ula-response/ula-response-getters.test.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Coöperatieve Rabobank U.A.
+ * Copyright 2020 Coöperatieve Rabobank U.A.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,13 +15,14 @@
  */
 
 import { assert } from 'chai'
-import { UlaResponse } from '../../../src'
+import { UlaError, UlaResponse } from '../../../src'
 
 const testData = {
-  statusCode: 200,
+  statusCode: 'some-status-code',
   body: {
     canBeAnything: 'anything'
-  } as object
+  },
+  error: new UlaError('some-status-code', 'Something went wrong!')
 }
 
 describe('UlaResponse getters', function () {
@@ -35,7 +36,17 @@ describe('UlaResponse getters', function () {
     assert.strictEqual(sut.body, testData.body)
   })
 
-  it('should flatten an object using JSON.stringify()', () => {
-    assert.strictEqual(JSON.stringify(sut), `{"statusCode":200,"body":{"canBeAnything":"anything"}}`)
+  it('should return an unchanged error', () => {
+    assert.deepStrictEqual(sut.error, testData.error)
+  })
+
+  it('should flatten an object (with error) using JSON.stringify()', () => {
+    const sutError = sut.error as Error
+    assert.strictEqual(JSON.stringify(sut), `{"statusCode":"some-status-code","body":{"canBeAnything":"anything"},"error":{"message":"Something went wrong!","stack":${JSON.stringify(sutError.stack)}}}`)
+  })
+
+  it('should flatten an object (without error) using JSON.stringify()', () => {
+    const sutWithoutError = new UlaResponse({ statusCode: testData.statusCode, body: testData.body })
+    assert.strictEqual(JSON.stringify(sutWithoutError), `{"statusCode":"some-status-code","body":{"canBeAnything":"anything"}}`)
   })
 })